+++ Engine/Controls/Options/ResetSettingsControl.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using JocysCom.ClassLibrary.Controls;
-using System;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Windows;
@@ -19,12 +18,14 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls.Options
 				return;
 		}
 
-
-
+		private void ResetAllSettingsButton_Click(object sender, RoutedEventArgs e)
+		{
+			SettingsSourceManager.ResetAllSettings(true);
+		}
 
 		private void ResetApplicationSettingsButton_Click(object sender, RoutedEventArgs e)
 		{
-			if (!AppHelper.AllowReset("Application Settings, but not Services and Models"))
+			if (!AppHelper.AllowReset("Application Settings"))
 				return;
 			SettingsSourceManager.ResetAppSettings();
 		}
@@ -43,12 +44,12 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls.Options
 			SettingsSourceManager.ResetTemplates();
 		}
 
-		private void ResetPromptingButton_Click(object sender, RoutedEventArgs e)
+		private void ResetPromptsButton_Click(object sender, RoutedEventArgs e)
 		{
-			if (!AppHelper.AllowReset("Prompting Templates"))
+			if (!AppHelper.AllowReset("Promps"))
 				return;
 			SettingsSourceManager.ResetPrompts();
-			Global.PromptItems.Save();
+			Global.Prompts.Save();
 			Global.TriggerPromptingUpdated();
 		}
 
@@ -66,62 +67,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls.Options
 		/// </summary>
 		private void ResetUIButton_Click(object sender, RoutedEventArgs e)
 		{
-			var items = Global.AppSettings.PanelSettingsList.ToArray();
-			foreach (var item in items)
-				ClassLibrary.Runtime.Attributes.ResetPropertiesToDefault(item);
-			var ps = Global.AppSettings.StartPosition;
-			if (!Global.IsVsExtension)
-			{
-				var window = ControlsHelper.GetParent<Window>(this);
-				//var pixRect = PositionSettings.GetPixelsBoundaryRectangle(this);
-				//var pixRectWin = PositionSettings.GetPixelsBoundaryRectangle(window);
-				var w = Math.Max((double)WindowWidthUpDown.Value, window.MinWidth);
-				var h = Math.Max((double)WindowHeightUpDown.Value, window.MinHeight);
-				var content = (FrameworkElement)window.Content;
-				// Get space taken by the window borders.
-				var wSpace = window.ActualWidth - content.ActualWidth;
-				var hSpace = window.ActualHeight - content.ActualHeight;
-				//var tPad = pixRect.Top - pixRectWin.Top;
-				//var lPad = pixRect.Left - pixRectWin.Left;
-				//var padPoint = new Point(tPad, lPad);
-				var size = new Size(w + wSpace, h + hSpace);
-				var point = new Point(window.Left, window.Top);
-				var newSize = PositionSettings.ConvertToDiu(size);
-				var newPoint = PositionSettings.ConvertToDiu(point);
-				//var newPadPoint = PositionSettings.ConvertToDiu(padPoint);
-				ps.Left = (int)(newPoint.X / 2 / 3 / 5) * 2 * 3 * 5;
-				ps.Top = (int)(newPoint.Y / 2 / 3 / 5) * 2 * 3 * 5;
-				ps.Width = newSize.Width;
-				ps.Height = newSize.Height;
-				ps.LoadPosition(window);
-			}
-		}
-
-		/// <summary>
-		/// Adjusts the provided dimension to the nearest perfect size for screenshots, 
-		/// meeting the criteria of divisibility by 2, 3, 4, and 10.
-		/// </summary>
-		/// <param name="value">The original size of the screenshot dimension 
-		/// (width or height) to be adjusted.</param>
-		/// <returns>The adjusted size, meeting the criteria of being a multiple of 2, 3, 4, and 10 
-		/// for optimal resizing quality.</returns>
-		public static int AdjustForScreenshot(int value)
-		{
-			// The LCM of 2, 3, 4, and 10 to ensure scaling and quality criteria
-			const int perfectDivisor = 60;
-			// If the value already meets the perfect criteria then return.
-			if (value % perfectDivisor == 0)
-				return value;
-			// Calculate the nearest higher multiple of 60
-			int adjustedValue = ((value / perfectDivisor) + 1) * perfectDivisor;
-			return adjustedValue;
-		}
-
-		private void AdjustUIButton_Click(object sender, RoutedEventArgs e)
-		{
-			WindowWidthUpDown.Value = AdjustForScreenshot((int)WindowWidthUpDown.Value);
-			WindowHeightUpDown.Value = AdjustForScreenshot((int)WindowHeightUpDown.Value);
-			ResetUIButton_Click(null, null);
+			SettingsSourceManager.ResetUI();
 		}
 
 		private void ResetListsButton_Click(object sender, RoutedEventArgs e)