+++ Engine/Controls/SettingsListFileControl.xaml.cs
@@ -474,22 +474,73 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 
 		private void CopyButton_Click(object sender, RoutedEventArgs e)
 		{
-			var item = MainDataGrid.SelectedItems.Cast<ISettingsListFileItem>().FirstOrDefault();
-			if (item == null)
+			try
+			{
+				var selectedItems = MainDataGrid.SelectedItems.Cast<object>().ToArray();
+				if (!selectedItems.Any())
+					return;
+				// Create array for serialization.
+				var itemType = SourceItems.GetType().GenericTypeArguments[0];
+				var items = Array.CreateInstance(itemType, selectedItems.Length);
+				Array.Copy(selectedItems, items, items.Length);
+
+				// Serialize the items to a XML string.
+				var text = items.Length == 1
+					? Serializer.SerializeToXmlString(selectedItems[0], null, true)
+					: Serializer.SerializeToXmlString(items, null, true);
+				var files = new System.Collections.Specialized.StringCollection();
+				if (selectedItems[0] is ISettingsListFileItem)
+				{
+					foreach (var item in items)
+					{
+						var fi = (ISettingsListFileItem)item;
+						// Determine the file name
+						var fileName = $"{fi.Name}.xml";
+						// Write the serialized text to a file
+						var filePath = System.IO.Path.Combine(System.IO.Path.GetTempPath(), fileName);
+						System.IO.File.WriteAllText(filePath, text);
+						files.Add(filePath);
+					}
+				}
+				// Create a DataObject to hold both text and file drop list
+				var dataObject = new DataObject();
+				dataObject.SetText(text);
+				if (files.Count > 0)
+					dataObject.SetFileDropList(files);
+				// Place the dataObject on the clipboard
+				Clipboard.SetDataObject(dataObject, true);
+			}
+			catch (Exception ex)
+			{
+				Global.ShowError(ex.Message);
 				return;
-			var text = Serializer.SerializeToXmlString(item, null, true);
-			Clipboard.SetText(text);
+			}
 		}
 
 		private void PasteButton_Click(object sender, RoutedEventArgs e)
 		{
 			try
 			{
-				var xml = Clipboard.GetText();
+				var xml = Clipboard.GetText()?.Trim();
 				var itemType = SourceItems.GetType().GenericTypeArguments[0];
-				var item = (ISettingsListFileItem)Serializer.DeserializeFromXmlString(xml, itemType);
-				AppHelper.FixName(item, SourceItems);
-				InsertItem(item);
+				var isArray = xml.StartsWith($"<{nameof(Array)}");
+				if (isArray)
+				{
+					var arrayType = Array.CreateInstance(itemType, 0).GetType();
+					var items = Serializer.DeserializeFromXmlString(xml, arrayType) as Array;
+					foreach (var item in items)
+					{
+						var sfi = item as ISettingsListFileItem;
+						AppHelper.FixName(sfi, SourceItems);
+						InsertItem(sfi);
+					}
+				}
+				else
+				{
+					var item = (ISettingsListFileItem)Serializer.DeserializeFromXmlString(xml, itemType);
+					AppHelper.FixName(item, SourceItems);
+					InsertItem(item);
+				}
 			}
 			catch (Exception ex)
 			{